Changed build system
Now the cli code and the vm code are separated in the vm and cli folders.
When building the cli with the `all` target, first a static library of
the vm is compiled and the cli code is linked against it.
Furthermore, the user has now the option of building only the vm as
a static and dynamic library via the `libs` target (useful for embedding
the vm in another application).