Corrected a bug in compund assignment compilation
With previous implementation the ast incurred in a double free which
created some nasty bug. Current implementation uses a new ast node for
reppresenting a compound assignment, and only during compilation the
node gets expanded in its full form.
The implementation proceeds this way:
	a += b -> CompAssNode(operator = +, left = a, right = b)
during compilation this will be expanded to:
	AssNode(left = a, BinaryNode(operator = +, left = a, right = b))

then this new expanded ast gets fed into the normal compileAssignment
function.
These two new nodes are created on the stack in order to avoid the
double free problem that existed in previous implementation.