Added anonymous functions and changed def keyword to fun
Anonymous functions can now be created with the syntax:
	fun (**args**) **statements** end

Anonymous functions are expressions and they bind less tightly than a
postfix expression, so to call a newly created anonymous functions
parenthesis must be provided: (fun (**args**) **statements** end)().

Being expressions, anonymous functions can be assigned to variables or
passed as arguments. This simplifies the usage of common programming patterns,
such as a function taking a predicate to apply over its argument:
	fun forEach(iter, predicate)
		for var e in iter do
			predicate(e)
		end
	end

this function can now be called with an anonymous function as predicate
all in one expression:
	forEach([1, 2, 3], fun (e) print(e) end)

Note that the syntax for defining named functions,
fun name(**args**) **statements** end, is equivalent to:
	var name = fun (**args**) **statements** end
with the only difference being that the former syntax assigns "name" as
the name of the function at compile time, while the latter assigns
"anon:LineNo". This means that stacktraces and other debug output will
generally be more readable using named functions.