Project structure refactoring
the builtin folder has now been moved inside the vm folder, as the
source files in it are compiled to the libblang library.

Completely decoupled libblang from the parser, that now is compiled in a
separate library, libblparse. This neatly separates the parser from the
rest of the vm, giving a clear interface for the user to use the parser
if he wants.
For example the cli uses the parser interface to identify a correctly
formed expression and, if so, it adds a print statement to print the
result of such expression.