Fixed a bug in anonymous function creation
- Added a smarter way to resolve an anonymous function's name (this is
kinda oxymoronic). Basically, during compilation, the compiler looks if
the anonymous function is assigned to a variable (e.g. var a = fun ...
or  a = fun ...) and, if it is, it assigns the name of the identifier to
the function. This will work even on attribute assignments (o.name =
fun ...). If no sensible name can be found, then the function is named
anon:<line>, where line is the line in which it was defined.