Fixed problem in generation of atuomatic opcode names
The array containing the opcode names was generated without a comma
inbetween entries, causing mingw to compile it as a big string with all
the names one after another (with no NUL termination of entries) instead
of an array of const char *. This caused problems in disassembleIstr that
treats OpcodeName as an array. This bug strangely manifest only on mingw-w64,
while the code compiles OpcodeName as an array of const char * even without
commas on other compilers.