Updated README.md and tracked .jsr.h files
The README.md file has been updated to reflect the changes in the
language and in preparation of the creation of the language
documentation site.
Furthermore, jsr.h files are new tracked in the git repository so users
that do not need to modify jsr files can skip the dependency on the
python interpreter during builds.