Extend functionality of `super` to bound-methods
Can now bind an instance to a `super` version of a method. This is
achieved by using the same syntax used for normal bound-methods while we
are inside a method of a child class:
	var bound = super.methodToBind

Calling `bound` will result in the invocation of the super-class's
`methodToBind` with the instance bound at creation time as a receiver
(i.e. the instace assigned to `this` in the method where `bound` was
created).

Just like super-method calls, we can omit the name when we want to
bind a super-method with the same name as the current one:
	class A2 is A1
		fun methodToBind()
			// These two are equivalent
			var bound1 = super
			var bound2 = super.methodToBind

			// further code...
		end
	end

Other changes:
 - Extended curly call notation to `super` calls:
	super.method{'a' : 1, 'b' : 2}
 - Changed some functions of vm.c in order to reduce the shuffling of
   values inside the stack when calling overloaded operators.
 - Changed implementation of some API functions due to the change above
   in order to mantain same behaviour