API changes: Changed behaviour of jsrEquals
Now the function returns true in case the two values at slot1 and slot2
are equal and false in the case they're not or if an exception is thrown
by __eq__.
This differs from the previous behaviour in which it acted like a normal
function call, leaving the result/exception on top of the stack and
signaling the success or failure by the returning a boolean.
This change has been made because the __eq__ overload shouldn't really throw
exceptions in correctly designed code. But, if such a situation arises,  we
need a way of solving it gracefully in the API. We do that by
discarding the exception and assuming the two values are not equal.