Extended language standard library with Exception.getStacktrace()
Exception.getStacktrace() will return a formatted string representing
the stacktrace of the Exception. The c API has been extended too, with a
function named 'jsrGetStacktrace(JStarVM *Vm, int slot) thath places
the formatted string at the top of the stack.
With this functions the user is able to get the error message and handle
it the way it wants (log it, store it on file, etc...), without being forced
to either handle the exception or to print it immediately to stderr with
printStacktrace.

Other changes:
 - Modified some function comments to better explain their behaviour