Massive restructuring of standard library
Moved all essential functionalities out of module `sys` into `core`, so
that the language can be compiled without the `sys` module.
Futhermore, `system` and `exec` have been moved from `core` to `sys`, as
they are specific to the system.
All of the changes will be documented upon the release ov version 1.3