__core__ is not implicitly imported by other modules anymore
Instead a second look-up is performed on the core module in case a
global variable is not found on the current one.
This change makes it impossible to use core module functions
transitively via another imported module. This is probabily the
correct thing to do as it eliminates possible confusion regarding the
location of the core functions, which should always be used like `global
scope` functions, and not as members of another module.
Unfortunately, to implement this behaviour, we incur in a double look-up
of a hash table (one for the current module, and then on the core if it
fails) and this costs us some performance when running code that
extensively uses global variables.