Sped up method calls by changing the semantics of method calls
Now methods are searched before fields when attempting to call methods
on an object, speeding up method calls significantly as we do not require
a double look-up every time.

This optimization comes with some drawbacks, a notable one being the
impossibility of fields shadowing a method on normal method calls.
One must instead be explicit and save the method in a variable before
calling it if shadowing is desired:
	var shadowed = obj.field // This will return a field if it shadows an object with the same name
	shadowed()