Fix bug in exception propragation
Now the compiler emits code to set the exception cause to null instead
of the exception itself after it has been handled. The code of
OP_END_HANDLER has been adjusted accordingly to check the exception
cause when determening if an exception has been handled or not.
This fixes a bug that arises while executing ensure handlers due to a
return with a `null` argument.