Fix bug in table constructor
Table constructor would crash the interpreter if the provided argument
was an empty Table (more precisely a Table with a capacity of 0), due to
how its capacity is stored internally. Now before proceeding with the
copy we check if the input table has at least 1 non-tombstone entry,
and therefore a capacity greater than 0.