Add `bind` method to `Function` class
`Function.bind` creates a new bound-method that replaces the original
receiver of the given function/native/other bound method with the one
provided as argument.

NOTE: You cannot rebind a bound method that has as a receiver a built-in
class. This is done in order to prevent crashes when calling built-in methods
that expect a bespoke object to be located in slot 0 (the `this`
variable).