Decoupled import system from the core jstar library
An interface has been defined in `jstar.h` that permits the embedders to
define a custom import policy.

This has some major advantages: for example the reduction of complexity
(and thus size) of the core jstar library, as well as the removal of
platform dependent functions for loading shared library extensions from
the core library.

All the previous code for importing modules and shared library
extensions has been moved inside the cli application.