Setup API stack during import callback
Setup the api stack also during import callback.
The rationale for this is that an import callback is very similar to a
native function call, in that it could reentrantly call the VM and
execute arbitrary bytecode. In order to make the operation on the stack
consistent in how they work with native functions, we need to setup the
api stack to the current position (the current stack pointer of the VM).

This will correctly mantain the invariants checked by assertions that
popping past the current native boundary will result in an underflow
assertion being triggered. This'll provide more clear error messages and
render complex stack operation debbugging more simple.