Change some API (jstar.h) functions
Now `jsrPushNative`, jsrSetGlobal` and `jsrGetGlobal` raise an `ImportException`
if the specified module could not be found.
In the case of `jsrPushNative` and `jsrSetGlobal` this means that their
return value changed from `void` to `bool` in order to signal to clients
the fail state of the call.

This is a breaking change, part of the `2.0` release