Add `importPaths` to `core` module
This variable is intended to be used as a hint of paths the embedder
should look into resolve imports. Can be customized by both the
embedder (via the J* API) and the language user (by appending/inserting/deleting
into the `importPaths` global list).

The embedder is in no way obliged to respect the paths reported into
this variable during import resolution, especially when the import logic
is not in any way tied to a filesystem.
Nonetheless, if the embeder is in fact looking for import files on a
filesystem, it is reccomended to take this variable under consideration.