Fix `unterminated-string-initialization` warning introduced with GCC 15
This warns for initialization of non-nul terminated string using a
string literal in `serialize.c`. This is exactly what we want, as we use
this to check for a magic header inside compiled bytecode files, not as
a string.
To fix the issue the simplest thing is to move away from the literal syntax,
and instead exeplicitely initialize the buffer as an array.
This is a bummer, but probably the best solution to continue supporting
older GCC version without doing version checks to disable the warning.